fix: move execute*Sync functions to lib/sync/ to comply with Next.js 16 route export rules
Made-with: Cursor
